python/requirement.txt
absl-py==2.2.1
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.6.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
joblib==1.4.2
keras==2.15.0
kiwisolver==1.4.8
libclang==18.1.1
load-dotenv==0.1.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
ml-dtypes==0.2.0
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.6.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.14.1
